+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+### INI.A306
+- Changed Cargo Loader Light Handling on Freighter
+- Toggle Refuel Cover and Panel during Refueling
+- Optimized Refuel-Sync (closer to internal prog. Refuel)
+- Improved Engine Cover Removal
+
+### SYN.A220
+- Added door sync for 1L & cargo doors
+- Added option to toggle ExtPwr fault protection
+- Updated channel definitions to allow mute functionality (aircraft v.1.0.9a or above required)
+
 ## [0c543bd] - 2026-09-13
 
 ### PMDG.B777